Use floor() instead of floorf().
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@19749 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSScroller.m
@@ -970,7 +970,7 @@ static NSColor *scrollBarColor = nil;
 
 	  /* calc knob's position */
 	  knobPosition = _floatValue * (slotHeight - knobHeight);
-	  knobPosition = (float)floorf(knobPosition);
+	  knobPosition = floor(knobPosition);
 
 
 	  /* calc actual position */
